#19 authority

+++ routes/api.php
@@ -29,43 +29,40 @@ Route::group(
     function () {
 
         // api users
-        Route::get('/users', [UserController::class, 'index']);
-        Route::post('/users', [UserController::class, 'store']);
-        Route::get('/users/{id}', [UserController::class, 'show']);
-        Route::put('/users/{id}', [UserController::class, 'update']);
-        Route::delete('/users/{id}', [UserController::class, 'destroy']);
+        Route::get('/users', [UserController::class, 'index'])->middleware('permission:GET_USER');
+        Route::post('/users', [UserController::class, 'store'])->middleware('permission:CREATE_USER');
+        Route::get('/users/{id}', [UserController::class, 'show'])->middleware('permission:GET_USER_DETAIL');
+        Route::put('/users/{id}', [UserController::class, 'update'])->middleware('permission:PUT_USER');
+        Route::delete('/users/{id}', [UserController::class, 'destroy'])->middleware('permission:DELETE_USER');
 
         // api gắn/cập nhật roles cho users
-        Route::put('/users/{id}/assign-roles', [UserController::class, 'assignRoles']);
+        Route::put('/users/{id}/assign-roles', [UserController::class, 'assignRoles'])->middleware('permission:PUT_ASSIGN_ROLE');
 
         // api gắn permissions cho roles
-        Route::post('/roles/{id}/assign-permissions', [RoleController::class, 'assignPermissions']);
+        Route::post('/roles/{id}/assign-permissions', [RoleController::class, 'assignPermissions'])->middleware('permission:POST_ASSIGN_PERMISSION');
 
         // api produts
-        Route::get('/products', [ProductController::class, 'index'])
-            ->middleware('permission:GET_PRODUCT');
-        Route::post('/products', [ProductController::class, 'store'])
-            ->middleware('permission:CREATE_PRODUCT');;
-        Route::get('/products/{id}', [ProductController::class, 'show'])
-            ->middleware('permission:GET_PRODUCT_DETAIL');;
-        Route::put('/products/{id}', [ProductController::class, 'update'])
-            ->middleware('permission:UPDATE_PRODUCT');;
-        Route::delete('/products/{id}', [ProductController::class, 'destroy'])
-            ->middleware('permission:DELETE_PRODUCT');;
+        Route::get('/products', [ProductController::class, 'index'])->middleware('permission:GET_PRODUCT');
+        Route::post('/products', [ProductController::class, 'store'])->middleware('permission:CREATE_PRODUCT');
+        Route::get('/products/{id}', [ProductController::class, 'show'])->middleware('permission:GET_PRODUCT_DETAIL');
+        Route::put('/products/{id}', [ProductController::class, 'update'])->middleware('permission:PUT_PRODUCT');
+        Route::delete('/products/{id}', [ProductController::class, 'destroy'])->middleware('permission:DELETE_PRODUCT');
 
         // api roles
-        Route::get('/roles', [RoleController::class, 'index']);
-        Route::post('/roles', [RoleController::class, 'store']);
-        Route::get('/roles/{id}', [RoleController::class, 'show']);
-        Route::put('/roles/{id}', [RoleController::class, 'update']);
-        Route::delete('/roles/{id}', [RoleController::class, 'destroy']);
+        Route::get('/roles', [RoleController::class, 'index'])->middleware('permission:GET_ROLE');
+        Route::post('/roles', [RoleController::class, 'store'])->middleware('permission:CREATE_ROLE');
+        Route::get('/roles/{id}', [RoleController::class, 'show'])->middleware('permission:GET_ROLE_DETAIL');
+        Route::put('/roles/{id}', [RoleController::class, 'update'])->middleware('permission:PUT_ROLE');;
+        Route::delete('/roles/{id}', [RoleController::class, 'destroy'])->middleware('permission:DELETE_ROLE');
+
 
         // api permissions
-        Route::get('/permissions', [PermissionController::class, 'index']);
-        Route::post('/permissions', [PermissionController::class, 'store']);
-        Route::get('/permissions/{id}', [PermissionController::class, 'show']);
-        Route::put('/permissions/{id}', [PermissionController::class, 'update']);
-        Route::delete('/permissions/{id}', [PermissionController::class, 'destroy']);
+        Route::get('/permissions', [PermissionController::class, 'index'])->middleware('permission:GET_PERMISSION');
+        Route::post('/permissions', [PermissionController::class, 'store'])->middleware('permission:CREATE_PERMISSION');
+        Route::get('/permissions/{id}', [PermissionController::class, 'show'])->middleware('permission:GET_PERMISSION_DETAIL');
+        Route::put('/permissions/{id}', [PermissionController::class, 'update'])->middleware('permission:PUT_PERMISSION');;
+        Route::delete('/permissions/{id}', [PermissionController::class, 'destroy'])->middleware('permission:DELETE_PERMISSION');
+
 
         // api auth
         Route::get('auth/profile', [AuthController::class, 'profile']);
